refactor(setup): extract model config inspection
